% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_switches.R
\name{extract_geometric_switches}
\alias{extract_geometric_switches}
\title{Extract Information on the number of geometric space switches from kolam.}
\usage{
extract_geometric_switches(sequence, sequential = FALSE)
}
\arguments{
\item{sequence}{The kolam drawing.}

\item{sequential}{A logical scalar. If \code{sequential} == FALSE, switches are counted for every not sequential transitional gesture. If \code{sequential} == TRUE, each counted transitional gesture equals 2 space switches (one before and one after).}
}
\value{
The total number of geometric switches.
}
\description{
\code{extract_geometric_switches} returns a count of the total number of geometric switches for a sequence.
}
\examples{
extract_geometric_switches(c( 'o1', 'o3', 'o4', 'o1', 'o1', 'o3', 'o4', 'o1', 'o1', 'o2', 'o1', 'o3', 'o4', 'o1', 'o1', 'o3', 'o4', 'o1', 'o1', 'o2',
'o1', 'o3', 'o4', 'o1', 'o1', 'o3', 'o4', 'o1', 'o1', 'o2', 'o1', 'o3', 'o4', 'o1', 'o1', 'o3', 'o4', 'o1', 'o1', 'o2'), sequential = FALSE)

}
