% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sequence.R
\name{extract_sequence}
\alias{extract_sequence}
\title{Extract the string of gestures from the untranformed sequence from the raw kolam data.}
\usage{
extract_sequence(untransformed_sequence)
}
\arguments{
\item{untransformed_sequence}{The untransformed sequence of gestures from the raw kolam data.}
}
\value{
A vector of strings. The string vector contains all the gestures of the sequence.
}
\description{
\code{extract_sequence} returns a vector of gestures.
}
\examples{
extract_sequence(untransformed_sequence = '[[O1 O1 O3 O1 O2 O1 O1 O3 O1 O2] [O2 O1 O3 O1 O1 O2 O1 O3 O1 O1] [O2 O1 O3 O1 O1 O2 O1 O3 O1 O1] [O2 O1 O3 O1 O1 O2 O1 O3 O1 O1]]')

}
