% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLoopInformation.R
\name{extractControlpoints}
\alias{extractControlpoints}
\title{Extract Information on control points and path from kolam.}
\usage{
extractControlpoints(kolamObject)
}
\arguments{
\item{kolamObject}{The kolam drawing.}
}
\value{
A list with control points and paths.
}
\description{
\code{extractControlpoints} returns a list with list entries controlpoints and path. The function enables the user to extract information from each kolam drawing.
}
\examples{
extractControlpoints(kolamObject)

}
