% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hornLeft.R
\name{h3lshortLong}
\alias{h3lshortLong}
\title{Return the control points of the bezier curve, path and starting information that would be used to draw the gesture h3lshortLong.}
\usage{
h3lshortLong(xStart, yStart, headingStart = 45, method = "bezier")
}
\arguments{
\item{xStart}{The start position of the x-coordinate. Can be a decimal number with .5 or whole number. But either x- and y- coordinate must have one whole number and a decimal number.}

\item{yStart}{The start position of the y-coordinate. Can be a decimal number with .5 or whole number.But either x- and y- coordinate must have one whole number and a decimal number.}

\item{headingStart}{The start orientation of the gesture. Valid headings or orientations are 45째, 135째, 225째 and 315째.}
}
\value{
A list with the starting information (x- and y-coordinate, initial orientation), the path and the control points.
}
\description{
\code{h3lshortLong} This function returns the points of the path and the exact control points that would be used to draw the series of line segments for the bezier curve. The control points capture the shape of the gesture. Horns are drawn on orthogonal gestures. Horns can be drawn starting from the long side or starting from the short side and this determines the curve, direction and the shape of the gesture. Furthermore, horn gestures are chiral, so there is a right and left version of the horn gestures.
}
\examples{
h3lshortLong(xStart = 0, yStart = 5.5, headingStart = 225)

}
