% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadKolam_Netlogo.R
\name{loadKolam}
\alias{loadKolam}
\title{Reload Saved Kolam Data}
\usage{
loadKolam(data)
}
\arguments{
\item{data}{The kolam data extracted from a .yaml file.}
}
\value{
A kolam object that contains a list of informaton.
\itemize{
  \item \code{sequence} The sequence of gestures for each loop.
  \item \code{df} A dataframe of starting information for all gestures in the dataframe.
  \item \code{nLoops} The total number of loops that comprise the kolam drawing.
  \item \code{nMatrix} Actual size the space (x- and y-coordinate) to draw the kolam.
  \item \code{pulli} Initial matrix of dots.
  \item \code{hash} A unique identifier for the kolam drawing.
  \item \code{initialization} A dataframe of starting information (x- and y-coordinates, initial heading) for each loop.
  \item \code{kolam_codes} Codes about the kolam drawing.
  \item \code{notes} Notes about the kolam drawing.
}
}
\description{
\code{loadKolam} Reloads the kolam object from a -yaml file to create a kolam object that contains a list of information.
}
\examples{
df <- loadKolam(df)

}
