% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalGestures.R
\name{o3}
\alias{o3}
\title{Return the control points of the bezier curve, path and starting information that would be used to draw the gesture O3.}
\usage{
o3(xStart, yStart, headingStart = 45, method = "bezier")
}
\arguments{
\item{xStart}{The start position of the x-coordinate. Can be a decimal number with .5 or whole number. But either x- and y- coordinate must have one whole number and a decimal number.}

\item{yStart}{The start position of the y-coordinate. Can be a decimal number with .5 or whole number.But either x- and y- coordinate must have one whole number and a decimal number.}

\item{headingStart}{The start orientation of the gesture. Valid headings or orientations are 45째, 135째, 225째 and 315째.}
}
\value{
A list with the starting information (x- and y-coordinate, initial orientation), the path and the control points.
}
\description{
\code{o3} This function returns the points of the path and the exact control points that would be used to draw the series of line segments for the bezier curve. The control points capture the shape of the gesture. The function draws an orthogonal gesture. A right turn and a left turn have different gesture vectors, different heading changes (+ and -), and different Bezier control points (switch x and y columns). If 0.5 is on a y-line, then an angle of 45 or 225 is a left turn, and an angle of 135 or 315 is a right turn. If 0.5 is on an x-line, then an angle of 45 or 225 is a right turn, and an angle of 135 or 315 is a left turn.
}
\examples{
o3(xStart = 0, yStart = 5.5, headingStart = 135)

}
