% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKolam.R
\name{plotKolam}
\alias{plotKolam}
\title{Plots a Kolam.}
\usage{
plotKolam(
  kolamObject,
  loop_col = TRUE,
  delay = NA,
  pulli = TRUE,
  arrow = TRUE,
  lwd = 2,
  cex = 1
)
}
\arguments{
\item{kolamObject}{The kolam object.}

\item{loop_col}{logical; if \code{TRUE} each loop is coloured in a different colour.}

\item{delay}{defaults to \code{NA}; suspends execution of R expressions for a specified time interval.}

\item{pulli}{logical; if \code{TRUE} the pulli dots in the matrix are not displayed and plotted.}

\item{arrow}{logical; if \code{TRUE} the arrows located at the starting position of each loop are not displayed and plotted.}

\item{lwd}{The line width, a positive number, defaulting to 2.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the  1}
}
\value{
A kolam object that contains a list of informaton.
\itemize{
  \item \code{sequence} The sequence of gestures for each loop.
  \item \code{df} A dataframe of starting information for all gestures in the dataframe.
  \item \code{nLoops} The total number of loops that comprise the kolam drawing.
  \item \code{nMatrix} Actual size the space (x- and y-coordinate) to draw the kolam.
  \item \code{pulli} Initial matrix of dots.
  \item \code{hash} A unique identifier for the kolam drawing.
  \item \code{initialization} A dataframe of starting information (x- and y-coordinates, initial heading) for each loop.
  \item \code{kolam_codes} Codes about the kolam drawing.
  \item \code{notes} Notes about the kolam drawing.
}
}
\description{
\code{plotKolam} Plots the kolam object after \code{\link{loadKolam}} extraction from a .yaml file to create a kolam object.
}
\examples{
plotKolam(df)

}
