% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchLoop.R
\name{plotLoop}
\alias{plotLoop}
\title{Plots one kolam loop.}
\usage{
plotLoop(
  gestures,
  xStart = 0,
  yStart = 0.5,
  headingStart = 45,
  delay = NA,
  add = FALSE,
  col = "black",
  windowMin = NA,
  windowMax = NA,
  Netlogo = TRUE,
  pulli = TRUE,
  arrow = TRUE,
  lwd = 2,
  cex = 1
)
}
\arguments{
\item{gestures}{The sequence vector of gestures.}

\item{xStart}{The start position of the x-coordinate.}

\item{yStart}{The start position of the y-coordinate.}

\item{headingStart}{The start orientation.}

\item{delay}{defaults to \code{NA}; suspends execution of R expressions for a specified time interval.}

\item{add}{logical; if \code{TRUE} the loop is plotted onto the previous plot.}

\item{col}{A character indicating the colour to be used to plot the loop.}

\item{windowMin}{The minimum value of the window.}

\item{Netlogo}{logical; if \code{TRUE} The headings/orientations are changed because the heading/orientations in this package are switched.}

\item{pulli}{logical; if \code{TRUE} the pulli dots in the matrix are not displayed and plotted.}

\item{arrow}{logical; if \code{TRUE} the arrows located at the starting position of each loop are not displayed and plotted.}

\item{lwd}{The line width, a positive number, defaulting to 2.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the  1}

\item{windowMaxThe}{maximum value of the window.}
}
\value{
Plot of a loop.
}
\description{
\code{plotLoop} Plots one loop of a kolam.
}
\examples{
plotLoop(c("o4", "o2", "o1", "o4", "o3", "o1", "o1", "o1", "o4", "o3", "o1", "o1", "o4", "o3", "o2", "o1"), delay = 0.0001, headingStart = 135)

}
