% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKolam_csv.R
\name{readKolam}
\alias{readKolam}
\title{Creates a list of multiple kolam objects from a .csv file.}
\usage{
readKolam(file = NA, data = NA, read_csv = FALSE)
}
\arguments{
\item{file}{Path to the .csv file.}

\item{data}{The dataframe in the R environment.}

\item{read_csv}{logical; if \code{TRUE}, the function \code{readKolam} reads the .csv file from the path given and then creates a list of all kolam objects.}
}
\value{
A kolam object that contains a list of informaton.
#' @examples
}
\description{
\code{readKolam} Reads in a .csv file with multiple kolams and then creates a list of all kolam objects found in the .csv file. Or creates a list of all kolam objects from a dataframe after the.csv file has been read in separately.
}
