% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{rotateDiagonalPoints}
\alias{rotateDiagonalPoints}
\title{Rotate diagonal gestures}
\usage{
rotateDiagonalPoints(headingStart, pathPoints)
}
\arguments{
\item{headingStart}{Only rotations in diagonal space are allowed. Valid rotations are: 90째, 180째, 270째 and 360째.}

\item{pathPoints}{A matrix of x- and y-coordinates with two columns.}
}
\value{
Returns the rotated matrix for diagonal gestures.
}
\description{
\code{rotateDiagonalPoints} rotates diagonal gestures in space and returns the rotated gestures.
}
\examples{
  gestureVector <- matrix(c(0.5, 0), ncol = 2) # unique to t4Left
  gestureVector <- rotateDiagonalPoints(180, gestureVector)

}
