% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{rotateHornPoints}
\alias{rotateHornPoints}
\title{Rotate special horn gestures}
\usage{
rotateHornPoints(headingStart, pathPoints)
}
\arguments{
\item{headingStart}{Only rotations in horn space are allowed. Valid rotations are: 45째, 135째, 225째 and 315째.}

\item{pathPoints}{A matrix of x- and y-coordinates with two columns.}
}
\value{
Returns the rotated matrix for horn gestures.
}
\description{
\code{rotateHornPoints} rotates horn gestures in space and returns the rotated gestures.
}
\examples{
gestureVector <- matrix(c(0, 0, -0.5, 0.5), ncol = 2, byrow = TRUE)
gestureVector <- rotateOrthogonalPoints(90, gestureVector)

}
