% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{rotateOrthogonalPoints}
\alias{rotateOrthogonalPoints}
\title{Rotate orthogonal gestures}
\usage{
rotateOrthogonalPoints(headingStart, pathPoints)
}
\arguments{
\item{headingStart}{Only rotations in orthogonal space are allowed. Valid rotations are: 45째, 135째, 225째 and 315째.}

\item{pathPoints}{A matrix of x- and y-coordinates with two columns.}
}
\value{
Returns the rotated matrix for orthogonal gestures.
}
\description{
\code{rotateOrthogonalPoints} rotates orthogonal gestures in space and returns the rotated gestures.
}
\examples{
gestureVector <- matrix(c(0.5, 0.5), ncol = 2)
gestureVector <- rotateOrthogonalPoints(45, gestureVector)

}
