% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchLoop.R
\name{sketchLoop}
\alias{sketchLoop}
\title{Sketches one kolam loop using a sequence of gestures.}
\usage{
sketchLoop(gestures, xStart = 0, yStart = 0.5, headingStart = 45)
}
\arguments{
\item{gestures}{The sequence vector of gestures.}

\item{xStart}{The start position of the x-coordinate.}

\item{yStart}{The start position of the y-coordinate.}

\item{headingStart}{The start orientation.}
}
\value{
A list of information on staring positions and headings, the control points and the path across all gestures in the sequence.
}
\description{
\code{sketchLoop} Plots one loop of a kolam.
}
\examples{
sketchLoop(c("o4", "o2", "o1", "o4", "o3", "o1", "o1", "o1", "o4", "o3", "o1", "o1", "o4", "o3", "o2", "o1"), xStart = 0, yStart = 0.5, headingStart = 135)

}
