% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonalOrthogonalTransitions.R
\name{t2Right}
\alias{t2Right}
\title{Return the control points of the bezier curve, path and starting information that would be used to draw the gesture t2Right.}
\usage{
t2Right(xStart, yStart, headingStart, method = "bezier")
}
\arguments{
\item{xStart}{The start position of the x-coordinate. Valid x-coordinates must be a number with decimal of 0.5.}

\item{yStart}{The start position of the y-coordinate. Valid y-coordinates must be a number with decimal of 0.5.}

\item{headingStart}{The start orientation of the gesture. Valid headings or orientations are 90째, 180째, 270째 and 360째.}
}
\value{
A list with the starting information (x- and y-coordinate, initial orientation), the path and the control points.
}
\description{
\code{t2Right} This function returns the points of the path and the exact control points that would be used to draw the series of line segments for the bezier curve. t2Right is a transitional gesture and allows to switch the geometric space from a diagonal gesture to an orthogonal gesture. Hence, the starting information must be in alignment with the rules for diagonal gestures. The control points capture the shape of the gesture. There is \code{t2Left} and \code{t2Right} because there are different ways to draw the t2 gesture, depending on the direction. Therefore, we have chirality.
}
\examples{
t2Right(xStart = 0.5, yStart = 1.5, headingStart = 90)

}
