% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalDiagonalTransitions.R
\name{t3}
\alias{t3}
\title{Return the control points of the bezier curve, path and starting information that would be used to draw the gesture T3.}
\usage{
t3(xStart, yStart, headingStart, method = "bezier")
}
\arguments{
\item{xStart}{The start position of the x-coordinate. Can be a decimal number with .5 or whole number. But either x- and y- coordinate must have one whole number and a decimal number.}

\item{yStart}{The start position of the y-coordinate. Can be a decimal number with .5 or whole number. But either x- and y- coordinate must have one whole number and a decimal number.}

\item{headingStart}{The start orientation of the gesture. Valid headings or orientations are 45째, 135째, 225째 and 315째.}
}
\value{
A list with the starting information (x- and y-coordinate, initial orientation), the path and the control points.
}
\description{
\code{t3} This function returns the points of the path and the exact control points that would be used to draw the series of line segments for the bezier curve. The control points capture the shape of the gesture. The function draws a transitional gesture. This transitional gesture allows to switch the geometric space from an orthogonal gesture to a diagonal gesture. Hence, the starting information must be in alignment with the rules for orthogonal gestures
}
\examples{
t3(xStart = 3.5, yStart = 5, headingStart = 225)

}
